A variety of 2D data loading performance improvements
Summary:
- read 2D geometry in new thread
- New global Line2DIntersectionManager object that is used where needed
- IODir has a hash map for fast id search
- Seis::SequentialReader will create CubeData for a single geomid only

Fixed a crash due to IODir::permRemove [devel]

The objects were removed without updating the hashmap objidmap_,
causing crashes in subsequent calls to IODir::get / IODir::indexOf

Restored loading of 2D geometries in IOM()

The filling of the geometries needs to stay at the end of IOMan::init.
This is essential to ensure that all applications that use project data, thus initialize IOM():
1- Will fill the geometry manager after having set the active survey
2- Will empty and reset the geometry manager on survey change
thus not only od_main, but also the main stand-alone apps, the batch programs and test programs.

Since the functionality is now moved to the Geometry manager, this module must provide to the General module a pointer
to the static function that needs to be executed.
Ensured that all batch programs will initialize the Geometry module. This is not done by the initialization of the Network
module which is independent of the Geometry module.

Fixed a crash when closing the application, especially for fast test_programs: The GeometryManager singleton is deleted
before the thread-work manager is deleted. Thus the Geom2DInit queue must be emptied when the GeometryManager is deleted,
and there is no need to call Thread::WorkManager::tm().shutdown, or attach to Thread::WorkManager::tm().isShuttingDown.

-- Include missing header file

-- Fixed mExpClass

Test Plan:
Data loading speed much, much faster. Intersection calculation still too slow.

The test programs have now a non-empty GeometryManager, and can convert line names to Pos::GeomID objects.
4 failing test programs work.
test_angle_computer still has a crash to be fixed, which does not occur when restoring the call to Survey::GMAdmin().fillGeometries
in IOMan::init

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: nanne, raman

Differential Revision: https://phabricator.dgbes.com/D17526