Fixed threading issues while changing surveys [devel]
Summary:
The four functions SurveyInfo::setSurveyName, SurveyInfo::curSurveyName, SetCurBaseDataDir and GetBaseDataDir must be using the same (static) thread lock to ensure that survey changes are done safely, especially when using the SurveyChanger class in a for loop.
Also the FileSpec object in the IOStream implementation of IOObj was having several GetDataDir calls. This is not correct: The value of GetDataDir will change when switching to another survey. Therefore the IOStream object must receive the current value of GetDataDir during the construction of this object. Passing it using a DBKey to the FileSpec class constructor. IOX and IOSubDir implementations of IOObj do not need it, but will still receive it.
I choose to store GetDataDir as a string in the FileSpec class because the DBKey and its SurveyDiskLocation are too dynamic, they may be calling SI(), which again will change its content when switching surveys.
Reworked Geom2DInit, to get the possibility to not compute the bendpoints if the geometry reading has been interrupted, and not compute the intersections of the bendpoints reading has been interrupted. This is done using the new SurveyInfo::rootDirChanged() global notifier. Note that it is not possible to attach this notification directly to Geom2DInit::closeQueueCB, as this will mostly likely cause a dead lock in either SurveyInfo::curSurveyName() or GetBaseDataDir().

Test Plan:
* All test programs run without error. These are the two importants:
test_surveychanger  (starts with a programmatically set project, and set uses several SurveyChanger objects). Should never be reading the current survey/data root from the user settings.
test_angle_computer (starts reading the initial survey from the user settings (if any), and then uses several SurveyChanger objects). Should only read the current survey/data root from the user settings are its beginning.
Added extra log output to check the order of operations.

* od_main starts correctly, can display data in the scene and data managers.
Tested switching to another project and coming back, looked at the extra log output.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17535