Added a warning when running older kernels [devel]
Summary:
Qt6 cannot capture the stdout of a QProcess on Linux kernel v4(.18.0), at least for the patch 372, and problably all patches below.
This corresponds to RHEL 8.0 to RHEL 8.6 (included). RHEL 8.7 provides kernel 4(.18.0)-425, which does not suffer from this issue.
Added a contextual warning the first time any MachineCommand is executed on the affected platforms.

[Jira: OD-5958]

Test Plan: tweaked the versiosn to test if the message would be printed as expected.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17537