Added cmake support for CMAKE_DEBUG_POSTFIX [devel]
cmake would automatically use that variable if set to set the filename of the link and runtime libraries.
Ensured that the applications and test programs would also be using that same postfix for their filename.
Ensured that the built-in 'test' cmake target would launch the test programs correctly.
Passing the cmake value into the Basic library by setting a compilation definition.

For the C++ API:
Added HasDebugPostFix() and GetDebugPostFix(). These two functions are
mostly internal to the Basic module.
Ensured that the loading of the plugins is working (API forced loading and using the ALO files).

Added GetODApplicationName, to be used at all places in the source code where a hard-coded application name is used.
Until these locations have been found and updated, the added lines in the root CMakeLists.txt file must remain commented.
This has already been done for the few test programs that needed it.

Testing: od_main starts correctly and all plugins are loaded. All test programs succeed in debug mode with the default postfix.