Refined the style/platform selection [devel]
Summary:
The style and platform settings of a Qt app can be set on the command line using the -style and -platform keywords.
The corresponding values, if set, should not be ignored by uiMain::preInit and uiMain::init, and take precedence over the eventual user settings and environment variables.
Also looks at the content of the 'QStyleSheet::factory':
On Windows it contains 'windowsvista', 'Windows', and 'Fusion' styles for both Qt5 and Qt6. Qt5 uses the 'windowsvista' style by default. Qt6 has the additional Windows11 style, which becomes the default on that platform.
On Unix is contains 'Windows' and 'Fusion', default 'Fusion' on Unix, 'Windows' on macOS.
None of the styles honor the dark mode with Qt5, but the styles 'Windows', 'Fusion' and 'Windows11' do honor the dark mode theme with Qt6.

Thus disabling the dark mode needs to dependent on the (user or hard-coded) style selection, and be ignored if explicitly set on the command line.

Do we want to go back to a hard-coded windowsvista style on Windows, when not set externally, to provide the sme look and feel as in v7?
Let's implement a user-defined style selection in another revision

Test Plan: Tested a variety of command line switches and environment variable. The styles are quite easy to tell apart, already on the plugin selection window.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17561