Added copying of the PDB files for dependent libraries [devel]
Summary:
The runtime DLLs of plugins where copied into the binaries folder, but not
there PDB files, which is not practical when debugging in MSVC.
Unfortunately this is no generator expression equivalent to TARGET_RUNTIME_DLLS
for the PDB files, hence needed to create a script, CopyPDBs.cmake, that
can be called with that same generator expression.
The script will receive from cmake the list of DLLs for which we want to copy
also the PDB files, and that list will be configuration dependent as it is
coming from a generator expression.
The actual implementation is quite simple: Replace the 'dll' extension by 'pdb',
copy the file is it exists and is coming from another directory than the destination (needed because some the input DLLs are OpendTect base libraries: Their PDB file is already in the destination folder).

Testing:
Remove all DLLS from bin/win64/Debug, rebuilt the solution: The PDB files from the Qt, OSG and HDF5 libraries are copied as expected.