Forwardding the graphics information to the System namespace
Summary:
Added System::graphicsInformation to hold a static IOPar that can be used
to read this information in modules that do not depend on uiOSG.
The set may be empty if either the information has not been fetched (uiOSG is not loaded),
or the platform does not provide any information (macOS).
One can build one uiGenInput per IOPar item, using the item key as label, and item value as value.

Needed to rework a bit how the information is gathered in uiOSG, and transformed into a message
when one needs to show a pop-up message: The contextual pop-up message at application startup needs to stay.
Also fixed a caching issue: If one has stored GL information from let's say an intel card, but starts
OpendTect with the nvidia card as the current GPU, the stored settings key 'dTect.Last GL info' should be removed
from the settings, it was not removed until now.

Test Plan:
Graphics information is displayed from the Utilities menu in the same current format.
System::graphicsInformation is not empty after the initialization of the uiOSG module.

Reviewers: nanne

Reviewed By: nanne

Subscribers: hadyan

Differential Revision: https://phabricator.dgbes.com/D17574