Fixed implementation of Pos::GeomID::is2D [devel]
Summary:
That function was returning true for Pos::Geom::udf(), which is not correct. is2D, is3D and isSynth should only return true for a valid Pos::GeomID.
This was probably unseed in v7 because Pos::GeomID::is2D was hardly used, which is no longer the case.

Test Plan: reverted a change made to uiCreate2DGrid::checkLineNames, which should not have been needed. Still working as expected.

Reviewers: nanne

Reviewed By: nanne

Subscribers: wayne, raman

Differential Revision: https://phabricator.dgbes.com/D17575