Fixed external Polygon display issue
Summary: While getting a Polygon/PolyLine from Shapefile or Petrel, we set the Type as well as the disp_.connect_ parameter. The setDefaultDispPars() should not overwrite it unless the type is Polygon and the disp_.connect_ still defaults to None.

Test Plan: displayed PointSets, Polygons, Polylines and Shapefiles. They all look good now.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17578