Fixed CheckHelpIDs script [devel]
The script was using all files in plugins/*/*, which is too inclusive.
Some of these files are binary files, like icon file, and should not be
taken into account.

When too many of these binary files have been included, the composite
file $tmpfile is seen as a binary file, resulting in all HelpIDs being
though as missing, when it is not the case.

Also reverted an old change, which was specific to the 6.6 branch.