Fixed destination in OD_WRITE_MODDEPS on macOS
The wrong destination in the build directory was used on macOS, when
copying the ModDeps.* file from the source directory to the build
directory. This prevented the ModDepMgr class from loading and
initializing the core modules on macOS, leading to many runtime errors.

Did not affect the releases, as the installation destination was good.