Fixed errors in System network implementations [devel]
The implementation of System::isAcceptable was not correct for the non-ipv4 case,
it was still returning true for an input Ipv4 address.

Also System::hostName was returning the wrong hostname when received the IPv4 of the local network adapter. On Windows it could return "host.docker.internal" when docker is installed, which is valid, but not the expected local hostname.