Ensure session restore works with od8 developers version [devel]
Summary:
Session files written by od8 development version could not be restored because the version
in the file is "2025.DEV". The "DEV" minor version was set to and udf(int). This
change traps undefined values read by iopar.cc in the minor and patch version and replaces them with 0.

Test Plan:
Tested on Linux:
- saved a session
- restored the session

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17612