Fixed failure to read markers from HDF5 wells
Summary: markers were not being read from HDF5 wells. This was because they were being written in double format, but while reading the data was requested in float format. This caused the chunk size incompatibilty with DataSetKey for markers(64-bit) and that of float(32-bit). So, to fix it, not while reading, we are going to request double instead of float.

Test Plan: tested by reading wells into Advance Well Manager.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17625