Added setSize/setCapacity to the ObjectSet class [devel]
Summary:
Similar to the implementation in the TypeSet class, with the difference that the setSize function cannot have a default T object to fill the vector.
Instead, the vector gets automatically filled with size nullptr, which is the preffered method, as these nullptr can then by replaced by the actual objects using the so-called replace function

Also added a multi-threaded ObjectSet deleter, similar to deepErase function, but on multi threads. Did not add it into the Basic module, to avoid introducing dependencies between the paralleltask and objectset classes.

Test Plan:
using two test programs to QC the implementation
The test_arraymath program will consume around 1GB of memory on windows (runtime very fast, below 1 sec)
The ObjectSetDeleter is running as expected correctly on all available threads, mostly because the ObjectSet::get and ObjectSet::replace functions do not contain any thread lock.

Reviewers: raman, nanne, arnaud

Reviewed By: arnaud

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com/D17648