Made the attribute log creation fault tolerant
Summary:
too many 'return false' in the for loop doing the work.
Converted them to 'continue', and accumulating the errors in a uiRetVal.
[Jira: OD-6056]

Test Plan:
using the project cited in the Jira issue, where one of the 402 wells would always
fail, because it is oustide of the coverage of the input seismic dataset.
Now all wells but this one well are processed and stored, and an error message showing the name
of the failing well is displayed.

Reviewers: arnaud

Subscribers: lloyd, wayne, raman

Differential Revision: https://phabricator.dgbes.com/D17659