Update number to string API [devel]
Made it more easy to format a BufferString and uiString according to
cformat specifications.
This should eventually allow to:
1- Remove all od_sprintf calls outside of the base string classes
2- Get a similar API for uiString objects
3- Print numbers using uiString objects in the current locale, instead
of the C-locale (disabled for now, can be enabled by setting the
environment variable OD_USE_LOCALE_NUMBERS).

Removed lot's of very old code from string2.cc. Aligned the default
arguments with the values from the Qt QString class.
It is questionable whether we should keep the toStringLim and
BufferString::setLim functions, as they will truncate some numbers if
the maximum value is set too low.