Add simple text progress reporting as an option to od_process_attrib
Summary:
Adds a simple progress reporting option for od_process_attrib that can be more easily parsed for the OXY attribute progress reporting.
I had to add the HiddenParams to the SequentialTask class because the constructor/destructor of the Executor class were defined in the header file and could not be
modified without impacting the ABI stability.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17644