Added arguments to uiTable::setValue functions [devel]
nrdec was far from being sufficient. Since the table is performing a
number to string conversion, one needs to be able to use the full set of
arguments of the toString functions.

Also adapted usage: Using 'f' format specifier for coordinates and depth
values, with the appropriate number of decimals.