Fixed restarting of batch jobs [devel]
Summary:
The program name writen in the IOPar file is configuration independent,
but the Batch::JobSpec::prognm_ object is not, it is the filename of the
application to be started, hence it is configuration dependent.

Thus in order to be reliably able to restart batch jobs, one needs to:
1- Add the debug postfix if applicable in Batch::JobSpec::usePar
2- Remove it in Batch::JobSpec::fillPar

Test Plan:
Defined and processed a similarity attribute in a debug build -> OK
Defined and processed a similarity attribute in a release build -> OK
Restarted a similarity job in the debug build -> OK
Restarted a similarity job in the release build -> OK
The parameter file is identical in these four cases (Re-starting a job
does re-write the parameter file).