Enabled auto-update of plane displays in the 3D scene [devel]
Summary:
This mechanism monitors changes made to the datasets registered in the OpendTect database,
to trigger a contextual action.
The only action implemented for now applies to inline, crossline, Z-slices, random line
and 2D seismic displays: The attribute layers which display stored datasets in the 3D scene(s)
will automatically be recalculated, hence their display will be updated if the file content
has changed (or anything else that will trigger IOM().implUpdated).

Test Plan:
Added a 15sec timer that triggers IOM().implUpdated,
displayed seismic on an inline, monitored that uiODApplMgr::getNewData
is called again when the timer is hit, for the selected MultiID.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17683