Moved getFileSize function to base transl class with IOObj* arg
Summary: This function should be in the base class so that it can be used elsewhere. The argument IOObj* is for now an optional one, as the function was moved to seis tarnslator where ioobj is not really required. The seis tarnslator has a Conn obbject which can give ioobj when required.

Test Plan: tested in seismic manager.

Reviewers: khushnood, arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17693