Removed warining pixmap in manager listbox, now coloring the entry
Summary: The QListWidgetItem supports just a single icon to be set. So setting the warning pixmap for a bad object would unset the icon for the object type. That is not good because now the user has no idea what type the entry is for. So, we are now painting the background red to show that there is something wrong with the entry.

Test Plan: tested in Seismic manager

Reviewers: khushnood, arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17697