Multiple fixes in the launchers [devel]
Summary:
The syntax used when calling create_target_launchers cmake function was incorrect,
and relied on cmake variables which no longer exist.
Made it work with proper generator expressions, which account for the
Debug vs. Release folder structure, and the presence or absence of a
debug postfix.
Also added the feature that sets the --datadir and parameter file
argument(s) for test programs that need it.
Finally, setting OD_USER_PLUGIN_DIR such that the Tut exercise points
automatically to the ODWork folder, and that the plugins are auto-loaded
without have to install the project.

Test Plan:
From visual studio, the test programs can now be running without have to
manually edit the target properties by adding --data and the full path
to the parameter file.
From visual studio, using the Tut exercise, it is now possible again to
start the 'od_main', od_process_attrib' and 'od_process_volume' targets,
and the tutorial plugins are auto-loaded as they should.