Porting changes from main to 8.0
Summary:
1) Made the labels and filepaths better in Edit Direct dat dialog
Edit direct data dialog now has labels based on the object it is linking to. Also, locations and files have been disambiguated as labels and specified in the dialog so that there is no confusion on file selection vs folder selection.

2) Moved getFileSize function to base transl class with IOObj* arg
This function should be in the base class so that it can be used elsewhere. The argument IOObj* is for now an optional one, as the function was moved to seis tarnslator where ioobj is not really required. The seis tarnslator has a Conn obbject which can give ioobj when required.

3)Removed warining pixmap in manager listbox, now coloring the entry
The QListWidgetItem supports just a single icon to be set. So setting the warning pixmap for a bad object would unset the icon for the object type. That is not good because now the user has no idea what type the entry is for. So, we are now painting the background red to show that there is something wrong with the entry.

Test Plan: tested in Seismic manager

Reviewers: khushnood, arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com/D17698