Multiple fixes in the launchers [devel]
Summary:
Ported some of the changes from the main branch, using the cmake variables
that exist in this branch.
Also added the feature that sets the --datadir and parameter file
argument(s) for test programs that need it.
Finally, setting OD_USER_PLUGIN_DIR such that the Tut exercise points
automatically to the ODWork folder, and that the plugins are auto-loaded
without have to install the project.

Test Plan:
From visual studio, the test programs can now be running without have to
manually edit the target properties by adding --datadir and the full path
to the parameter file.