Fix the kernel information layout
Summary: In the new information dialog, particularly on the Linux platform, the kernel information field is displayed incorrectly as it gets duplicated with the username information field. This revision resolves the issue by positioning the kernel information field below the username field.

Test Plan:
On a Linux platform, navigate to Utilities > Installation > Information.
Verify that the username and kernel information fields are displayed separately with distinct layouts.

Reviewers: hadyan, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17704