cmake bug fixes for the Tutorial [devel]
The logical in CreateLaunchers.cmake was incorrect for this branch, and reverted
some of the changes made previously.

The plugins Tut/uiTut should create ALO files when used in the tutorial exercise.
Setting the CMAKE FOLDERS to on
Setting a CMAKE_FOLDER for the launchers
Using the OpendTect SDK cmake variables to provide the list of launcher program names