Multiple fixes in the launchers [devel]
Summary:
Ported some of the changes from the main branch, using the cmake variables
that exist in this branch.
Also added the feature that sets the --datadir and parameter file
argument(s) for test programs that need it.
Setting OD_USER_PLUGIN_DIR such that the Tut exercise points
automatically to the ODWork folder, and that the plugins are auto-loaded
without having to install the project.
For the Tutorial:
The plugins Tut/uiTut should create ALO files when used in the tutorial exercise.
Setting the CMAKE FOLDERS to on
Setting a CMAKE_FOLDER for the launchers
Using the OpendTect SDK cmake variables to provide the list of launcher program names

Test Plan:
From visual studio, the test programs can now be run without having to
manually edit the target properties by adding --datadir and the full path
to the parameter file.
In the tutorial plugin, using MSVC, the tutorial plugin libraries are auto-loaded when executing the target od_main_console.