Fixed cmake error when using OD_INSTALL_DATADIR_MOD [devel]
Summary:
OD_INSTALL_DATADIR_MOD could not be used by an external plugins repository, if the OpendTect
SDK installation was read-only. It should write its files to the 'data' folder of that external
plugins repository, and not inside the SDK installation.
The new default location for writing these files is the CMAKE_BINARY_DIR, which thus points to the
external repository build directory, not the OpendTect SDK. This location can optionally be set
by providing an optional second argument to the macro OD_INSTALL_DATADIR_MOD.

To make it work, one needed to also change this implementation that reads these files from their
data folder. The three current static functions thus have to read the data folder of the OpendTect
SDK installation, and the data folder of the external repository build. This is done in the same
way as the plugin manager loads these foreign libraries, by reading the environment variable OD_USER_PLUGIN_DIR.

This is the correct implementation that reverts this incomplete commit:
https://github.com/OpendTect/OpendTect/commit/133bde820d363c55fa7ca63508933bea7def498e

Test Plan:
Configured the OpendTect project: The generated files remain created in this build's 'data' subfolder
Configured the Tutorial plugin exercise, by adding a cmake call to OD_INSTALL_DATADIR_MOD in both plugins.
Their generated files are now created in the corresponding build directory.
Started od_main in gdb, checked that the filepaths used to look for the InitXXXSel external function are valid for:
OpendTect modules
OpendTect plugins
External plugins.

Reviewers: wayne, raman

Reviewed By: raman

Subscribers: raman, nanne

Differential Revision: https://phabricator.dgbes.com/D17712