Fix python activation and internal python libs usage on macOS
This PR ensures that the python miniconda environment is properly activated on macOS.
It depends on OpendTect/odpy#4

It ensures that external python libraries like odpy, safety and odbind can be used and imported in this python environment at ease by including them in the PYTHONPATH variable.

It also ensures that external libraries like fontconfig, freetype and libpng are always available in new development builds.