Fixed the annotations for depth scene with diff display and storage unit
Summary: The grid and annotations in the 3D scene are now nice integers even when the display unit is different from storage unit in depth surveys. In doing so, Vis annotation's API has been cleaned up. To set a scale, just specifying the scale factor is not enough. The caller should also provide the scaled TrcKeyZSampling and the scale factor for that scaled sampling.

Test Plan: tested in depth survey, time survey with depth display in feet

Reviewers: khushnood, arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17731