Fix for mismatch between random line and TrcKeyPath
Summary: While moving the nodes of the random line display interactively, the Geometry::RandomLine object was updated and provided the input for the new DataPack being read, but the TrcKeyPath was not updated resulting in pError and data mismatch

Test Plan: No more perrors. The data displayed matches the geometry.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17744