Fixed several bugs for top/bot images
Summary:
When using the import button, the input field was left empty,
and the (option) item was still switched off. That was mostly because uiSurvTopBotImageGrp::newImageCB
did not create a new object if none was available.
Also the import window can be used to import several datasets, hence it must
have a trigger that can be used to forward the 'ready' operation, with the
corresponding MultiID, before leaving the acceptOK function of uiImportImageDlg.
Applied the same logic to uiEditImageDlg, in case we need it.
The MultiID was not saved in sessions, hence not correctly restored, fixed it as well

Replaced ImageDef::filename_ by a proper FileSpec object,
and replaced visBase::TopBotImage::filenm_ by an ImageDef object.
This makes the code cleaner, and allows saving relative filepaths within the definition file.

[Jira: OD-6163]

Test Plan:
Tested creation of new top/bot image datasets, their edition,
but also selecting image datasets that are already loaded in the project.
Top/bot images restored from legacy sessions are displayed in the 3D scene, but their dialog
shows an empty input.
That input field is now correctly field when restoring new session files, that contain
visBase::TopBotImage::odimageid_ instead of a filepath.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17756