Fixed display of time/frequency panels
Summary:
The zaxis and display settings of the time/frequency panels were not correct
for the spectral decomposition and CEEMD attributes.
The flatviewer ZDomain was not set from the FlatDataPack ZDomain, and the category of the
FlatDataPack was inappropriate.
This is because the AttributeEngine returns a (non-flat) datapack with a standard category.
One needs to be able to create a FlatDataPack with a category different from its source
datapack

Test Plan:
Spectral Decomposition and CEEMD time/frequency panels have now the time z-axis
correctly scaled and labeled, and the clipping settings are different from the standard 'Attribute.V'
datapack category.
Both attribute can however share the same "Time-Frequency spectrum" category.

Reviewers: khushnood

Reviewed By: khushnood

Subscribers: raman, nanne

Differential Revision: https://phabricator.dgbes.com/D17771