Added SeisIOObjInfo::getDataRange [devel]
Summary:
some seismic data implementations can directly provide the amplitude range,
thus added a function to retrieve it. This function MUST be per component of a dataset.
For instance a TFL dataset has completely different ranges in the three components.

If the translator cannot provide the range, it is retrieved from the dataset distribution
calculator.
This calculator was also not 'component' aware, thus added the corresponding implementation.

Test Plan:
tested from a test program. Exists stats are read and used, new per component stats
are added on top of the 'all components' distribution information.

Reviewers: khushnood

Reviewed By: khushnood

Subscribers: raman, nanne

Differential Revision: https://phabricator.dgbes.com/D17775