Added uiSeisSubSel::setToAll() [devel]
Summary:
uiSeis3DSubSel::clear() does not do anything, as it only calls uiObject::clear().
If one wants to reset that GUI, one must be able to call uiSeisSubSel::selfld_->setToAll(),
which actually resets this object.
Also fixed a typo in uiPosSubSel::setToAll, which now gives the expected behaviour:
If the selection is set to a range subselection, and 'None' is allowed, calling setToAll
should switch to the 'None' provider (and reset the internal TrcKeyZSampling, which was already done)

Test Plan:
Tested from the wavelet extraction GUI (temporary code, it does not have a reset button).
Will be used in plugins

Reviewers: khushnood

Reviewed By: khushnood

Subscribers: raman, nanne

Differential Revision: https://phabricator.dgbes.com/D17776