Fixed issue in initPluginClasses [devel]
Summary:
Batch programs that load their plugins directly, not using the ALO files,
will get problems with the initPluginClasses: This function is called before
entering the BatchProgram::doWork function, hence the plugin manager PIM() does not have
the entries for the libraries not listed in the ALO files.
An alternative to PIM().findData must be used for such libraries, and that is
to construct the path to the plugin library using either PIM().getAutoDir(true)
or PIM().getAutoDir(false).

Test Plan: All test batch programs work now.

Reviewers: khushnood, nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17780