Fixed incorrect trigger of IOM().implUpdated for 2D data [devel]
Summary:
The SeisTrc2DTranslator behaves strangely when being closed:
It has a valid connection, but the connection is 'Bad', because it points to a folder within the Seismics directory. A 'bad', or valid but closed connection will always return false to the forRead() function.
SeisTrcTranslator::writeBlock() does not return false, because it has nothing to write,
thus eventually IOM().implUpdated is trigger incorrectly.
There is probably room for more improvements.

[Jira: OD-6193]

Test Plan: No more crash, trigger is no longer hit during read operations.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17786