Option to enforce reload from disk in MultiWellReader
Summary:
multiwellreader calls Well:MGR().get(...) function to retrieve a well. As a result, if the well has already been read, then you always get one that was in the memory already. Sometimes, it is nice to reload from disk, i.e. read the well anew and discard any existing changes.

This revision allows the user to enforceReload of wells. Thus for the wells that have already been read, Well::MGR().reload(...) function is called, and the well is read from the disk.

Test Plan: tested in Advance Well Manager

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17795