Fixed crash when removing faults
Summary: This crash also affected faultsets and geobodies. probably Horizon2D as well, but didn't test with it. Basically, anything using polyline3D class. The destructor was unreffing its OSG::Node itself, but that shouldn't be the case. If the node is set as osg node, by calling setOSGNode function, the base visBase::DataObject class handles unreffing in it's destrucutor.

Test Plan: tested in 3D scene with faults and geobodies

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17877