Fixed issues with the contextual toolbar
Summary:
Two issues
1- Historical issue: When adding a new slice, its contextual toolbar is missing most of the items
This is because uiODDisplayTreeItem::init is calling uiVisPartServer::setSelObjectId before the
slice has a name, and before attaching tb->createnotifier to uiODDisplayTreeItem::addToToolBarCB
Other historical issue: In interact mode, picking on the non-selected displayed plane should makes it selected,
but the contextual toolbar was missing most of its items.

2- Qt6 specific issue: The order of the events listview_->mouseButtonPressed and listview_->selectionChanged
is different than in Qt5. In Qt6, listview_->mouseButtonPressed is emitted as soon as the mouse button is pressed,
whereas listview_->selectionChanged is only going to be emitted when the mouse button is released. Qt5 sends them in the reverse order.
Hence the actions uiTreeTopItem::anyButtonClickCB and uiTreeTopItem::selectionChanged are done in a different order.
Fix: In uiODSceneMgr::updateSelectedTreeItem, updateItemToolbar should only be called after
the 'updateSelXXX' calls in the for loop, as the 'selected' item is only correct for sure thereafter. Then the
order of the events in uiTreeTopItem no longer matters.

Test Plan:
Applied many actions:
- 'Add Default Data' on inline: Inline (425) should be selected, with all its default button in the toolbar
- Switch to the attribute displayed on this inline -> toolbar content should change
- Many times switched the selected item to scene, or inline number, or its data, QCed the toolbar content.
- Add crossline display, in interact mode, switch between the planes, QCed the toolbar content.
- Test tree selections with two scenes.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17879