Improved sorting of horizons
Summary: Instead of using just the number of positions above/below, it now uses the some of Z differences, which acts as a weighted count or volume above/below. The raw count is still used to find the number of crossings.

Test Plan: Tested sorting in a set of 20 complex horizons, merging with each other. The results are eaxctly as expected.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17868