Added protections to the geometry manager
Summary:
Sets of Geometry objects should be RefOjectSet<> objects, to automatically handle the ref/unref calls in a safe way. Adapted the manager, reader, writer and translators accordingly.

Remove the double book keeping of the Geometry3D singleton: Removed the copy in the SurveyInfo class, to keep it exclusively in the GeometryManager. Reworked the access functions accordingly. The main and more or less only access point should be Survey::Geometry3D::instance().
The SurveyInfo class does keep a Geometry3D object in a RefMan, for the work area (if set).

Fixed two crashes which occurs when the RefMan<CoordSystem> Coords::wgs84crsproj_ and RefMan<CoordSystem> SurveyInfo::coordsystem_ are emptied after the CRS library is unloaded: Since the library is unloaded, the application does not know how to destruct these objects, which is a derived implementation in the CRS plugin. This unloading happens automatically in function CloseBindings (not used by OpendTect). Thus needed to add the global function CleanupCRSPlugin
Also the CloseBindings function should start setting the is_exiting_ static flag: This allows preventing the GeomFileReader executor from starting if the executor is constructed when the application/the bindings are being closed:

Added protections to these three tasks:
Survey::GeomFileReader (executor)
BendPointFinder2DGeomSet (parallel task)
Line2DInterSectionFinder (parallel task)

All three will monitor 'IsExiting' at the start, during, and at the end of the execution, to stop their activity when the application is closed, or the bindings are closed.

Test Plan:
Tested from the test_odbinding application, using projects with a very small,. small, standard or very large number of geometries

- Small number of geometries: The GeomFileReader task is finished before the bindings are closed -> test OK
- Intermediate number of geometries (most challenging): The GeomFileReader task is started, but not finished while closing the bindings -> test OK
- Large number of geometries: The GeomFileReader task is not yet started by the time the bindings are closed -> test OK

From od_main:
- Started, closed the application
- Displayed 2D line geometries in 3D scene (without data), closed the application
- Displayed 2D line geometries with a displayed attribute in 3D scene, closed the application
- Created a new 2D line from a random line
- Created a grid of 2D lines from a 3D dataset

Reviewers: wayne

Reviewed By: wayne

Subscribers: raman, nanne

Differential Revision: https://phabricator.dgbes.com/D17884