Fixed copy within a Polygon in batch mode
Summary: In batch mode, SelData was always a RangeSelData (default) which amounted to the Polygon bounding box. Now, passing the SelData as part of SeisSorAccess::Setup fixes the problem.

Test Plan: Tested copy within a polygin for single and multicomponent datasets with one as well as all components selected. The output follows the polygon outline precisely.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17895