Fix for Provider destructor not being called
Summary:
Found AttribProvider destructor was not called because of circular reference in
the allexistingprov_ member. Replaced RefObjectSet with WeakPtrSet. Also in internalCreate
found new provider was added as a parent to its input, presume the input should be the parent.

Test Plan:
Tested on Linux using single attribute and attribute dependent on another attribute in the set.
With the debugger observed the Provider destructor being called.
Also applied the changes in the od8.0 branch and tested with wmplugins. The python runtime
started by the ExternalAttribute plugin observed to be shutdown correctly.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17905