Rework of the access to the resource and configuration files [devel]
Summary:
Added the possibility to define custom locations at runtime for the resource and configuration files: addCustomShareFolder
Renamed the 'data' folder name to the more standard 'share' folder name.
Reworked the old macros, added a set of functions that would allow finding the relevant resource/configuration files, knowing that they can be located in several folders. This allows:
- Finding a file or folder at the root of the 'share' folder(s)
- Finding a file or folder in a subfolder of the 'share' folder(s)
- Finding files/dir by pattern at the root and in a subfolder of the 'share' folder(s)
Also made a strong differentiation between the resource files (icons, various binary files, the BasicSurvey directory), to be accessed using the software location only, and the configuration files that can be also in the application setup and plugin directory locations. Checked the usage.

The legacy 'data' folder remains read if present to support legacy files from other vendors.

The 'user plugin dir' does then no longer need to be used outside of the Basic module (enum ODSetupLoc_UserPluginDirOnly).
The icon folders are reworked accordingly. The icons are now no longer copied into the build directory. Removed (actually moved) many icons that do not belong to the OpendTect SDK.

Change to the packaging: In the presence of a share/<package_name> folder at the installation location, the entire content of that folder goes automatically into the corresponding package by name. Mostly handy for the packaging of the external plugins.

Test Plan:
- cmake setup OK -> generated files are created at the expected location. icon files are not copied
- build OK
- runtime OK: No pErrMsg on missing icons or configuration files
- installation OK
- packaging OK (will need to update the QC files)

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17912